fix: split settings into multiple files
